Updated to svn r1299
- This includes Nico's changes to my burst mode code for dc1394
- Here is what he wrote:

Hi Greg,

I spend some time figuring out what was going on with burst mode in
the dc1394 adapter. It turned out that MMCore thought that the camera
was done acquiring right when it started.  The fact that you always
saw about the number of images that was in the buffer was because
those images were already in the buffer before you started.  Nenad
told/reminded me that MMCore checks whether the camera is done or not
using the Busy function.  Adding in the variable 'acquiring_' into
the Busy function (which actually lived in the header file to make it
even more obscure) fixes things.

Please be aware that the 'ResizeImageBuffer' function starts up the
whole image acquisition part of the camera.  I added another call to
the function in 'StartAcquisition' and also read away any frames that
might still be in the dc1394 buffer when starting.  I am not totally
sure that everything in ResizeImageBuffer is appropriate for what we
want it to do, but it suffices for now.

I don't really know what the set_multi_shot function does.  My camera
apparently does not support it, ignoring the error fixes things for me.

We should also put the conversion functions that are now in SnapImage
into seperate functions so that they can be caller from 'PushImage'
as well.

Thanks for all your work on this!

I committed my changes.

Best,

Nico